Fix the toppings price display.
The price breakdown either showed or hid the toppings for ALL of the items.
This was because the condition was set to check the variable $toppings that was
reassigned in each iteration, and ultimately held the value of the last one.

Fixed by changing the condition to check if toppings property of each object
($food->toppings) is empty, instead of checking the $toppings variable.